smallWidgetBlock added
