--- Corona Widget/Widget/Widget/Placeholder.swift
@@ -1,18 +1,0 @@
-//
-//  Placeholder.swift
-//  Corona Widget
-//
-//  Created by Aaryan Kothari on 24/07/20.
-//
-
-import SwiftUI
-
-// shown while data loads.
-
-struct Placeholder : View{
-    var body : some View {
-        Text("YOLO")
-    }
-}
-
-//END
