+++ Corona Widget/Widget/Widget/SmallWidget.swift
@@ -10,7 +10,10 @@ import SwiftUI
 //MARK: systemSmall Widget
 
 struct smallWidget : View {
+    // MARK - PROPERTIES
     var data : Countries
+    
+    // MARK - BODY
     var body : some View {
         VStack{
             HStack{
@@ -18,46 +21,23 @@ struct smallWidget : View {
                 Text(data.emoji())
                 Text(data.Country)
                 Spacer()
-            }.font(.headline)
+            } //: HSTACK
+            .font(.headline)
             .foregroundColor(.white)
             .padding(.top, 10)
             .padding(.all, 10)
             .background(Color.pink)
+            
             VStack(alignment:.leading){
                 smallWidgetBlock(type: .total, count: data.TotalConfirmed, color: .coronapink)
                 smallWidgetBlock(type: .recovered, count: data.TotalRecovered, color: .coronagreen)
                 smallWidgetBlock(type: .deaths, count: data.TotalRecovered, color: .coronagrey)
                 smallWidgetBlock(type: .active, count: data.totalActive(), color: .coronayellow)
-                
-            }.padding(.bottom, 15)
+            } //: VSTACK
+            .padding(.bottom, 15)
             Spacer()
-        }
-    }
-}
-
-struct smallWidgetBlock : View {
-    var type : coronaType
-    var count : Double
-    var color : Color
-    var body: some View {
-        HStack {
-            Image(type.image)
-                .resizable(capInsets: EdgeInsets(), resizingMode: .stretch)
-                .aspectRatio(contentMode: .fit)
-            
-            Text(type.rawValue)
-                .minimumScaleFactor(0.5)
-                .font(.system(size: 15))
-                .lineLimit(1)
-                .layoutPriority(1)
-            Text(count.cases)
-                .bold()
-                .minimumScaleFactor(0.5)
-                .lineLimit(1)
-                .layoutPriority(1)
-                .foregroundColor(color)
-        }
-    }
+        } //: VSTACK
+    } //: BODY
 }
 
 //END
