+++ Corona Widget/Widget/Widget/Widget.swift
@@ -8,19 +8,8 @@
 import WidgetKit
 import SwiftUI
 import Intents
-import MapKit
 
 struct DataProvider : TimelineProvider {
-    
-    func placeholder(in context: Context) -> CoronaData {
-        return CoronaData(Corona(Global: Global(NewConfirmed: 0, TotalConfirmed: 0, NewDeaths: 0, TotalDeaths: 0, NewRecovered: 0, TotalRecovered: 0), Countries: [Countries(Country: "India", CountryCode: "IN", Slug: "IN", NewConfirmed: 0, TotalConfirmed: 0, NewDeaths: 0, TotalDeaths: 0, NewRecovered: 0, TotalRecovered: 0, Date: "DATE")]))
-    }
-    
-    func getSnapshot(in context: Context, completion: @escaping (CoronaData) -> Void) {
-        coronaStore.fetch{ corona in
-            completion(CoronaData(corona))
-        }
-    }
     
     
     @ObservedObject var coronaStore = SessionStore()
@@ -36,46 +25,55 @@ struct DataProvider : TimelineProvider {
             completion(timeline)
         }
     }
-}
     
     
-    struct WidgetView : View{
-        var data : DataProvider.Entry
-        @Environment(\.widgetFamily) private var family
-        var body : some View {
-            Group {
-                switch family {
-                case .systemSmall:
-                    smallWidget(data: Country)
-                case .systemMedium:
-                    mediumWidget(data: Country)
-                case .systemLarge:
-                    largeWidget(data: data)
-                @unknown default:
-                    smallWidget(data: Country)
-                }
-            }
-        }
-        
-        var Country : Countries{
-            let country = CurrentCountry.county.rawValue
-            let countries = data.Countries
-            let mycountry = countries.filter { $0.CountryCode == country}
-            return mycountry.first!
+    func getSnapshot(in context: Context, completion: @escaping (CoronaData) -> Void) {
+        coronaStore.fetch{ corona in
+            completion(CoronaData(corona))
         }
     }
+    
+    func placeholder(in context: Context) -> CoronaData {
+        return CoronaData(Corona(Global: Global(NewConfirmed: 0, TotalConfirmed: 0, NewDeaths: 0, TotalDeaths: 0, NewRecovered: 0, TotalRecovered: 0), Countries: [Countries(Country: "India", CountryCode: "IN", Slug: "IN", NewConfirmed: 0, TotalConfirmed: 0, NewDeaths: 0, TotalDeaths: 0, NewRecovered: 0, TotalRecovered: 0, Date: "DATE")]))
+    }
+    
+}
 
 
-//@main
-//struct Config : Widget {
-//    var body: some WidgetConfiguration {
-//        StaticConfiguration(kind: "Widget", provider: DataProvider()) { (<#TimelineEntry#>) -> _ in
-//            <#code#>
-//        }
-//        StaticConfiguration(kind: "Widget", provider: DataProvider()) { data in
-//            WidgetView(data: data)
-//        }
-//        .supportedFamilies([.systemSmall,.systemMedium,.systemLarge])
-//        .description(Text("Current Time widget"))
-//    }
-//}
+struct WidgetView : View{
+    var data : DataProvider.Entry
+    @Environment(\.widgetFamily) private var family
+    var body : some View {
+        Group {
+            switch family {
+            case .systemSmall:
+                smallWidget(data: Country)
+            case .systemMedium:
+                mediumWidget(data: Country)
+            case .systemLarge:
+                largeWidget(data: data)
+            @unknown default:
+                smallWidget(data: Country)
+            }
+        }
+    }
+    
+    var Country : Countries{
+        let country = CurrentCountry.county.rawValue
+        let countries = data.Countries
+        let mycountry = countries.filter { $0.CountryCode == country}
+        return mycountry.first!
+    }
+}
+
+
+@main
+struct Config : Widget {
+    var body: some WidgetConfiguration {
+        StaticConfiguration(kind: "Widget", provider: DataProvider()) { data in
+            WidgetView(data: data)
+        }
+        .supportedFamilies([.systemSmall,.systemMedium,.systemLarge])
+        .description(Text("Current Time widget"))
+    }
+}
